Lootable Chest & Parry
Implemented chests with random loot inside and open animations. Also implemented a parry system with a riposte attack inflicting critical damage. Also implemented all the proper animations and improved the stamina system so that it works properly with all the player movements and prevents him to do certain activities while it's completely drained.