Dungeon & VFX
Implemented the dungeon blockout and reworked enemy AI for the dungeon.
Implemented the boss event as well as its UI for HP and name.
Implemented VFX for potions and a new one for healing.
Reworked some lighting and adjusted some post processing values.
Implemented cel-shading for most materials in game.
Implemented new mask and its materials and textures.
Implemented materials and textures for the golem enemy.
Implemented new tree asset and its textures for better performance.
Removed grass.
Improved loot from chests so that it supports loot tables for each type of items.