Bombs & Damage Over Time
Implemented a blood splat vfx when the player and enemies get hit.
Implemented a bomb consumable, animation and vfx.
Refactored the damage system to add compatibility for other types of damage implementation.
Implemented an ID system to identify friendly fire and avoid enemies with the same ID to detect the player and attack him and vice-versa.